+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [Unreleased]
+## [0.0.25] - 2019-05-29
 ### Added
 - `pf-dropdown` components can now inject the dropdown menu inside any element, with the `append-to` prop, to circumvent issues with clipped containers (e.g. `overflow:hidden`).
 - Addeed `pf-drawer-group-action` functional component to use inside the new `pf-drawer-group` `#actions` slot.
@@ -265,7 +265,8 @@ disabled, as per PatternFly design guidelines.
 - `pf-toolbar` component
 - `pf-utilization-bar-chart` component
 
-[Unreleased]: https://github.com/mtorromeo/vue-patternfly/compare/v0.0.24...HEAD
+[Unreleased]: https://github.com/mtorromeo/vue-patternfly/compare/v0.0.25...HEAD
+[0.0.25]: https://github.com/mtorromeo/vue-patternfly/compare/v0.0.24...v0.0.25
 [0.0.24]: https://github.com/mtorromeo/vue-patternfly/compare/v0.0.23...v0.0.24
 [0.0.23]: https://github.com/mtorromeo/vue-patternfly/compare/v0.0.22...v0.0.23
 [0.0.22]: https://github.com/mtorromeo/vue-patternfly/compare/v0.0.21...v0.0.22